# Conflicts:
#	.github/workflows/ci.yml
#	build.gradle
#	docker-compose.test.yml
#	src/main/java/com/example/taste/domain/notification/aop/EventAop.java
#	src/main/resources/application.properties
#	src/test/java/com/example/taste/domain/board/service/BoardServiceTest.java
#	src/test/java/com/example/taste/domain/pk/service/PkServiceTest.java
#	src/test/java/com/example/taste/domain/store/service/StoreBucketServiceTest.java
#	src/test/java/com/example/taste/fixtures/BoardFixture.java

+++ src/main/java/com/example/taste/domain/user/scheduler/UserScheduler.java
@@ -20,7 +20,7 @@ public class UserScheduler {
 	@Transactional
 	@Scheduled(cron = "0 0 0 1 * *")
 	public void resetPostingCnt() {
-		int updatedCount = userRepository.resetPostingCnt();
+		long updatedCount = userRepository.resetPostingCnt();
 		log.info("[UserScheduler] 총 {}명 유저 포스팅 횟수 초기화 완료", updatedCount);
 	}
 }
